Rename `bom` module to `bom_reference`
