+++ cyclonedx-bom/src/specs/common/bom_reference.rs
